Add files via upload
Large volumes of news articles are received everyday from various sources. To simplify downstream processing tasks, one of the first steps is to properly categorize news into well defined categories. In this problem, we have been given a set of news articles and a set of categories. The task is to develop a model which will be able to predict the category of unseen news.